extend player from new actor class
